Add a simple set of types for TCP keepalive
The keepalive algorithm typically used is to:

- wait until no packets have been received for a while (Linux has 2 hours)
- transmit up to `probes` probes with an `interval` between them (Linux
  defaults to 9 probes 75 seconds apart)
- if a response is received then the connection is alive!
- if no response is received then the connection is considered dead.

The types represent the `configuration` (initial timeout, intervals,
number of probes) and a mutable `state`. The client of the Keepalive module
should call `next` with the `configuration`, current `state` and the number
of nanoseconds since a packet is received. The response is a pair of
`action * state` where the `action` is one of:

- `Wait`: wait for more nanoseconds
- `SendProbe`: send a probe
- `Close`: consider the connection closed

Signed-off-by: David Scott <dave@recoil.org>